Edit entity jpa, repo and migration Liquibase change yml to xml

+++ src/main/java/com/huesosi/insuranceapi/repository/AgentDepartmentRepo.java
@@ -3,5 +3,5 @@ package com.huesosi.insuranceapi.repository;
 import com.huesosi.insuranceapi.persistance.jpa.AgentDepartment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgentDepartmentRepo extends JpaRepository<AgentDepartment, Integer> {
+public interface AgentDepartmentRepo extends JpaRepository<AgentDepartment, Long> {
 }
