+++ src/main/java/com/huesosi/insuranceapi/repository/ContractorTypeRepo.java
@@ -3,5 +3,5 @@ package com.huesosi.insuranceapi.repository;
 import com.huesosi.insuranceapi.persistance.jpa.ContractorType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContractorTypeRepo extends JpaRepository<ContractorType, Integer> {
+public interface ContractorTypeRepo extends JpaRepository<ContractorType, Long> {
 }
