+++ src/main/java/com/huesosi/insuranceapi/repository/InsuranceRiskRepo.java
@@ -3,5 +3,5 @@ package com.huesosi.insuranceapi.repository;
 import com.huesosi.insuranceapi.persistance.jpa.InsuranceRisk;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InsuranceRiskRepo extends JpaRepository<InsuranceRisk, Integer> {
+public interface InsuranceRiskRepo extends JpaRepository<InsuranceRisk, Long> {
 }
