+++ src/main/java/com/huesosi/insuranceapi/repository/MetadataRepo.java
@@ -3,5 +3,5 @@ package com.huesosi.insuranceapi.repository;
 import com.huesosi.insuranceapi.persistance.jpa.Metadata;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MetadataRepo extends JpaRepository<Metadata, Integer> {
+public interface MetadataRepo extends JpaRepository<Metadata, Long> {
 }
