+++ src/main/java/com/huesosi/insuranceapi/repository/ProductTypeRepo.java
@@ -3,5 +3,5 @@ package com.huesosi.insuranceapi.repository;
 import com.huesosi.insuranceapi.persistance.jpa.ProductType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductTypeRepo extends JpaRepository<ProductType, Integer> {
+public interface ProductTypeRepo extends JpaRepository<ProductType, Long> {
 }
