+++ src/main/java/com/huesosi/insuranceapi/repository/StatusContractRepo.java
@@ -3,5 +3,5 @@ package com.huesosi.insuranceapi.repository;
 import com.huesosi.insuranceapi.persistance.jpa.StatusContract;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusContractRepo extends JpaRepository<StatusContract, Integer> {
+public interface StatusContractRepo extends JpaRepository<StatusContract, Long> {
 }
